refactor(stable): update test scripts and enhance main process structure
- Updated the test script in package.json to include a new test file for image data URL handling.
- Refactored the main process to improve state management for signaling, replacing individual variables with a structured state object.
- Removed unused functions and interfaces related to screenshot handling and input capabilities to streamline the codebase.
- Added a new cache refresh feature in the settings page to allow users to manually refresh cached game data.

This commit enhances maintainability and introduces new functionality for better user experience.
