+++ opennow-stable/src/renderer/src/gfn/webrtcClient.ts
@@ -39,132 +39,33 @@ import {
 } from "./sdp";
 import { MicrophoneManager, type MicState, type MicStateChange } from "./microphoneManager";
 
-interface OfferSettings {
-  codec: VideoCodec;
-  colorQuality: ColorQuality;
-  resolution: string;
-  fps: number;
-  maxBitrateKbps: number;
-}
-
-interface RiInputCapabilities {
-  partialReliableThresholdMs: number | null;
-  hidDeviceMask: number;
-  enablePartiallyReliableTransferGamepad: number;
-  enablePartiallyReliableTransferHid: number;
-}
-
-interface DualRumbleEffectOptions {
-  startDelay: 0;
-  duration: number;
-  weakMagnitude: number;
-  strongMagnitude: number;
-}
-
-interface GamepadHapticActuatorLike {
-  readonly type?: string;
-  playEffect(effectType: "dual-rumble", options: DualRumbleEffectOptions): Promise<unknown>;
-}
-
-interface LegacyGamepadHapticActuatorLike {
-  pulse(value: number, duration: number): Promise<unknown>;
-}
-
-type GamepadWithOptionalHaptics = Gamepad & {
-  readonly vibrationActuator?: GamepadHapticActuatorLike | null;
-  readonly hapticActuators?: readonly (LegacyGamepadHapticActuatorLike | null | undefined)[] | null;
-};
-
-interface GamepadRumbleApi {
-  playEffectActuator: GamepadHapticActuatorLike | null;
-  pulseActuator: LegacyGamepadHapticActuatorLike | null;
-}
-
-interface ConnectedRumbleGamepad {
-  index: number;
-  gamepad: Gamepad;
-  api: GamepadRumbleApi | null;
-}
-
-function hevcPreferredProfileId(colorQuality: ColorQuality): 1 | 2 {
-  // 10-bit modes should prefer HEVC Main10 profile-id=2.
-  return colorQuality.startsWith("10bit") ? 2 : 1;
-}
-
-export interface StreamDiagnostics {
-  // Connection state
-  connectionState: RTCPeerConnectionState | "closed";
-  inputReady: boolean;
-  connectedGamepads: number;
-
-  // Video stats
-  resolution: string;
-  codec: string;
-  isHdr: boolean;
-  bitrateKbps: number;
-  decodeFps: number;
-  renderFps: number;
-
-  // Network stats
-  packetsLost: number;
-  packetsReceived: number;
-  packetLossPercent: number;
-  jitterMs: number;
-  rttMs: number;
-
-  // Frame counters
-  framesReceived: number;
-  framesDecoded: number;
-  framesDropped: number;
-
-  // Timing
-  decodeTimeMs: number;
-  renderTimeMs: number;
-  jitterBufferDelayMs: number;
-
-  // Input channel pressure
-  inputQueueBufferedBytes: number;
-  inputQueuePeakBufferedBytes: number;
-  partiallyReliableInputQueueBufferedBytes: number;
-  partiallyReliableInputQueuePeakBufferedBytes: number;
-  inputQueueDropCount: number;
-  inputQueueMaxSchedulingDelayMs: number;
-  partiallyReliableInputOpen: boolean;
-  mouseMoveTransport: "reliable" | "partially_reliable";
-  mouseFlushIntervalMs: number;
-  mousePacketsPerSecond: number;
-  mouseResidualMagnitude: number;
-  mouseAdaptiveFlushActive: boolean;
-
-  lagReason: StreamLagReason;
-  lagReasonDetail: string;
-
-  // System info
-  gpuType: string;
-  serverRegion: string;
-
-  // Decoder recovery status
-  decoderPressureActive: boolean;
-  decoderRecoveryAttempts: number;
-  decoderRecoveryAction: string;
-
-  // Microphone state
-  micState: MicState;
-  micEnabled: boolean;
-}
-
-export type StreamLagReason =
-  | "unknown"
-  | "stable"
-  | "network"
-  | "decoder"
-  | "input_backpressure"
-  | "render";
-
-export interface StreamTimeWarning {
-  code: 1 | 2 | 3;
-  secondsLeft?: number;
-}
+import type { StreamDiagnostics, StreamLagReason, StreamTimeWarning } from "./webrtc/streamDiagnosticsTypes";
+export type { StreamDiagnostics, StreamLagReason, StreamTimeWarning };
+import {
+  chooseAdaptiveMouseFlushInterval,
+  quantizeMouseDeltaWithResidual,
+  MouseDeltaFilter,
+  type AdaptiveMouseFlushDecisionParams,
+} from "./webrtc/mouseFlush";
+export { chooseAdaptiveMouseFlushInterval, quantizeMouseDeltaWithResidual };
+export type { AdaptiveMouseFlushDecisionParams };
+import {
+  hevcPreferredProfileId,
+  timestampUs,
+  parseRiInputCapabilities,
+  clampRumbleMagnitude,
+  isXboxLikeGamepad,
+  getGamepadRumbleApi,
+  parseResolution,
+  toRtcIceServers,
+  toBytes,
+  detectGpuType,
+  normalizeCodecName,
+  type OfferSettings,
+  type RiInputCapabilities,
+  type ConnectedRumbleGamepad,
+  type GamepadRumbleApi,
+} from "./webrtc/clientUtils";
 
 interface ClientOptions {
   videoElement: HTMLVideoElement;
@@ -184,349 +85,6 @@ interface ClientOptions {
   onTimeWarning?: (warning: StreamTimeWarning) => void;
   onMicStateChange?: (state: MicStateChange) => void;
 }
-
-function timestampUs(sourceTimestampMs?: number): bigint {
-  const base =
-    typeof sourceTimestampMs === "number" && Number.isFinite(sourceTimestampMs) && sourceTimestampMs >= 0
-      ? sourceTimestampMs
-      : performance.now();
-  return BigInt(Math.floor(base * 1000));
-}
-
-function parsePartialReliableThresholdMs(sdp: string): number | null {
-  const match = sdp.match(/a=ri\.partialReliableThresholdMs:(\d+)/i);
-  if (!match?.[1]) {
-    return null;
-  }
-  const parsed = Number.parseInt(match[1], 10);
-  if (!Number.isFinite(parsed) || parsed <= 0) {
-    return null;
-  }
-  return Math.max(1, Math.min(5000, parsed));
-}
-
-function parseRiIntegerAttribute(sdp: string, attribute: string, fallback: number): number {
-  const escapedAttribute = attribute.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-  const match = sdp.match(new RegExp(`a=${escapedAttribute}:([^\\r\\n]+)`, "i"));
-  const raw = match?.[1]?.trim();
-  if (!raw) {
-    return fallback;
-  }
-  const normalized = raw.toLowerCase();
-  const parsed = normalized.startsWith("0x")
-    ? Number.parseInt(normalized.slice(2), 16)
-    : Number.parseInt(normalized, 10);
-  return Number.isFinite(parsed) ? parsed : fallback;
-}
-
-function parseRiInputCapabilities(sdp: string): RiInputCapabilities {
-  return {
-    partialReliableThresholdMs: parsePartialReliableThresholdMs(sdp),
-    hidDeviceMask: parseRiIntegerAttribute(sdp, "ri.hidDeviceMask", PARTIALLY_RELIABLE_HID_DEVICE_MASK_ALL),
-    enablePartiallyReliableTransferGamepad: parseRiIntegerAttribute(
-      sdp,
-      "ri.enablePartiallyReliableTransferGamepad",
-      PARTIALLY_RELIABLE_GAMEPAD_MASK_ALL,
-    ),
-    enablePartiallyReliableTransferHid: parseRiIntegerAttribute(
-      sdp,
-      "ri.enablePartiallyReliableTransferHid",
-      PARTIALLY_RELIABLE_HID_DEVICE_MASK_ALL,
-    ),
-  };
-}
-
-function clampRumbleMagnitude(value: number): number {
-  if (!Number.isFinite(value)) {
-    return 0;
-  }
-  return Math.max(0, Math.min(1, value));
-}
-
-function isXboxLikeGamepad(gamepad: Gamepad): boolean {
-  return /xbox|xinput/i.test(gamepad.id);
-}
-
-function getGamepadRumbleApi(gamepad: Gamepad): GamepadRumbleApi | null {
-  const hapticGamepad = gamepad as GamepadWithOptionalHaptics;
-  const playEffectActuator = hapticGamepad.vibrationActuator;
-  const pulseActuator = hapticGamepad.hapticActuators?.[0];
-  const api: GamepadRumbleApi = {
-    playEffectActuator: playEffectActuator && typeof playEffectActuator.playEffect === "function"
-      ? playEffectActuator
-      : null,
-    pulseActuator: pulseActuator && typeof pulseActuator.pulse === "function"
-      ? pulseActuator
-      : null,
-  };
-  return api.playEffectActuator || api.pulseActuator ? api : null;
-}
-
-export interface AdaptiveMouseFlushDecisionParams {
-  baseIntervalMs: number;
-  currentIntervalMs: number;
-  reliableBufferedAmount: number;
-  schedulingDelayMs: number;
-  canUsePartiallyReliableMouse: boolean;
-  backpressureThresholdBytes: number;
-  minIntervalMs: number;
-  maxIntervalMs: number;
-}
-
-export function chooseAdaptiveMouseFlushInterval(params: AdaptiveMouseFlushDecisionParams): number {
-  const boundedBase = Math.max(params.minIntervalMs, Math.min(params.maxIntervalMs, params.baseIntervalMs));
-  const boundedCurrent = Math.max(params.minIntervalMs, Math.min(params.maxIntervalMs, params.currentIntervalMs));
-  if (!params.canUsePartiallyReliableMouse) {
-    return boundedBase;
-  }
-
-  const highPressure =
-    params.reliableBufferedAmount >= params.backpressureThresholdBytes / 2
-    || params.schedulingDelayMs >= 4;
-  if (highPressure) {
-    return Math.max(boundedBase, Math.min(params.maxIntervalMs, boundedCurrent + 2));
-  }
-
-  const lowPressure = params.reliableBufferedAmount <= 4096 && params.schedulingDelayMs <= 1;
-  if (lowPressure) {
-    return Math.max(params.minIntervalMs, boundedCurrent - 1);
-  }
-
-  if (boundedCurrent > boundedBase) {
-    return Math.max(boundedBase, boundedCurrent - 1);
-  }
-  if (boundedCurrent < boundedBase) {
-    return Math.min(boundedBase, boundedCurrent + 1);
-  }
-  return boundedCurrent;
-}
-
-export function quantizeMouseDeltaWithResidual(accumulatedDelta: number): { send: number; residual: number } {
-  const send = Math.round(accumulatedDelta);
-  return {
-    send,
-    residual: accumulatedDelta - send,
-  };
-}
-
-class MouseDeltaFilter {
-  private x = 0;
-  private y = 0;
-  private lastTsMs = 0;
-  private velocityX = 0;
-  private velocityY = 0;
-  private rejectedX = 0;
-  private rejectedY = 0;
-  private pendingX = 0;
-  private pendingY = 0;
-  private sawZero = false;
-  private relaxedForRawInput = false;
-
-  public setRelaxedForRawInput(value: boolean): void {
-    this.relaxedForRawInput = value;
-  }
-
-  public getX(): number {
-    return this.x;
-  }
-
-  public getY(): number {
-    return this.y;
-  }
-
-  public reset(): void {
-    this.x = 0;
-    this.y = 0;
-    this.lastTsMs = 0;
-    this.velocityX = 0;
-    this.velocityY = 0;
-    this.rejectedX = 0;
-    this.rejectedY = 0;
-    this.pendingX = 0;
-    this.pendingY = 0;
-    this.sawZero = false;
-  }
-
-  public update(dx: number, dy: number, tsMs: number): boolean {
-    if (dx === 0 && dy === 0) {
-      if (this.sawZero) {
-        this.pendingX = 0;
-        this.pendingY = 0;
-      } else {
-        this.sawZero = true;
-      }
-      return false;
-    }
-
-    this.sawZero = false;
-    if (this.pendingX === 0 && this.pendingY === 0) {
-      if (tsMs < this.lastTsMs) {
-        this.pendingX = dx;
-        this.pendingY = dy;
-        return false;
-      }
-    } else {
-      dx += this.pendingX;
-      dy += this.pendingY;
-      this.pendingX = 0;
-      this.pendingY = 0;
-    }
-
-    const dot = dx * this.x + dy * this.y;
-    const magIncoming = dx * dx + dy * dy;
-    const magPrev = this.x * this.x + this.y * this.y;
-    let accept = true;
-
-    const dtMs = tsMs - this.lastTsMs;
-    const directionReversalCosineThreshold = this.relaxedForRawInput ? 0.89 : 0.81;
-    if (dtMs < 0.95 && dot < 0 && magPrev !== 0 && dot * dot > directionReversalCosineThreshold * magIncoming * magPrev) {
-      const ratio = Math.sqrt(magIncoming) / Math.sqrt(magPrev);
-      let distToInt = Math.abs(ratio - Math.trunc(ratio));
-      if (distToInt > 0.5) {
-        distToInt = 1 - distToInt;
-      }
-      const intRatioRejectThreshold = this.relaxedForRawInput ? 0.07 : 0.1;
-      if (distToInt < intRatioRejectThreshold) {
-        accept = false;
-      }
-    }
-
-    const diffX = dx - this.x;
-    const diffY = dy - this.y;
-    const diffMag = diffX * diffX + diffY * diffY;
-
-    if (accept) {
-      const scale = 1 + 0.1 * Math.max(1, Math.min(16, dtMs));
-      const vx2 = 2 * scale * Math.abs(this.velocityX);
-      const vy2 = 2 * scale * Math.abs(this.velocityY);
-      const threshold = Math.max(this.relaxedForRawInput ? 9800 : 8100, vx2 * vx2 + vy2 * vy2);
-      accept = diffMag < threshold;
-      if (!accept && (this.rejectedX !== 0 || this.rejectedY !== 0)) {
-        const rx = dx - this.rejectedX;
-        const ry = dy - this.rejectedY;
-        accept = rx * rx + ry * ry < threshold;
-      }
-    }
-
-    if (accept) {
-      this.velocityX = 0.4 * this.velocityX + 0.6 * diffX;
-      this.velocityY = 0.4 * this.velocityY + 0.6 * diffY;
-      this.x = dx;
-      this.y = dy;
-      this.lastTsMs = tsMs;
-      this.rejectedX = 0;
-      this.rejectedY = 0;
-      return true;
-    }
-
-    this.rejectedX = dx;
-    this.rejectedY = dy;
-    return false;
-  }
-}
-
-function parseResolution(resolution: string): { width: number; height: number } {
-  const [rawWidth, rawHeight] = resolution.split("x");
-  const width = Number.parseInt(rawWidth ?? "", 10);
-  const height = Number.parseInt(rawHeight ?? "", 10);
-
-  if (!Number.isFinite(width) || !Number.isFinite(height) || width <= 0 || height <= 0) {
-    return { width: 1920, height: 1080 };
-  }
-
-  return { width, height };
-}
-
-function toRtcIceServers(iceServers: IceServer[]): RTCIceServer[] {
-  return iceServers.map((server) => ({
-    urls: server.urls,
-    username: server.username,
-    credential: server.credential,
-  }));
-}
-
-async function toBytes(data: string | Blob | ArrayBuffer): Promise<Uint8Array> {
-  if (typeof data === "string") {
-    return new TextEncoder().encode(data);
-  }
-  if (data instanceof ArrayBuffer) {
-    return new Uint8Array(data);
-  }
-  const arrayBuffer = await data.arrayBuffer();
-  return new Uint8Array(arrayBuffer);
-}
-
-/**
- * Detect GPU type using browser APIs
- * Uses WebGL renderer string to identify GPU vendor/model
- */
-function detectGpuType(): string {
-  try {
-    const canvas = document.createElement("canvas");
-    const gl = canvas.getContext("webgl2") || canvas.getContext("webgl");
-    if (!gl) {
-      return "Unknown";
-    }
-
-    const debugInfo = gl.getExtension("WEBGL_debug_renderer_info");
-    if (debugInfo) {
-      const vendor = gl.getParameter(debugInfo.UNMASKED_VENDOR_WEBGL);
-      const renderer = gl.getParameter(debugInfo.UNMASKED_RENDERER_WEBGL);
-
-      // Clean up renderer string - extract main GPU name
-      let gpuName = renderer;
-
-      // Remove common prefixes/suffixes for cleaner display
-      gpuName = gpuName
-        .replace(/\(R\)/g, "")
-        .replace(/\(TM\)/g, "")
-        .replace(/NVIDIA /i, "")
-        .replace(/AMD /i, "")
-        .replace(/Intel /i, "")
-        .replace(/Microsoft Corporation - /i, "")
-        .replace(/D3D12 /i, "")
-        .replace(/Direct3D11 /i, "")
-        .replace(/OpenGL Engine/i, "")
-        .trim();
-
-      // Limit length
-      if (gpuName.length > 30) {
-        gpuName = gpuName.substring(0, 27) + "...";
-      }
-
-      return gpuName || vendor || "Unknown";
-    }
-    return "Unknown";
-  } catch {
-    return "Unknown";
-  }
-}
-
-/**
- * Extract codec name from codecId string (e.g., "VP09" -> "VP9", "AV1X" -> "AV1")
- */
-function normalizeCodecName(codecId: string): string {
-  const upper = codecId.toUpperCase();
-
-  if (upper.startsWith("H264") || upper === "H264") {
-    return "H264";
-  }
-  if (upper.startsWith("H265") || upper === "H265" || upper.startsWith("HEVC")) {
-    return "H265";
-  }
-  if (upper.startsWith("AV1")) {
-    return "AV1";
-  }
-  if (upper.startsWith("VP9") || upper.startsWith("VP09")) {
-    return "VP9";
-  }
-  if (upper.startsWith("VP8")) {
-    return "VP8";
-  }
-
-  return codecId;
-}
-
 export class GfnWebRtcClient {
   private readonly videoStream = new MediaStream();
   private readonly audioStream = new MediaStream();
