+++ opennow-stable/src/renderer/src/styles.css
@@ -5577,109 +5577,6 @@ button.game-card-store-chip.owned.active:hover {
   gap: 10px;
 }
 
-.queue-ad-fallback {
-  position: relative;
-  width: 100%;
-  min-height: 180px;
-  overflow: hidden;
-  background: linear-gradient(135deg, rgba(11, 20, 14, 0.98), rgba(4, 8, 6, 0.98));
-}
-
-.queue-ad-fallback--compact {
-  min-height: 160px;
-}
-
-.queue-ad-fallback-bg {
-  position: absolute;
-  inset: 0;
-}
-
-.queue-ad-fallback-image {
-  position: absolute;
-  inset: 0;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  opacity: 0.28;
-  filter: saturate(1.1) blur(0px);
-}
-
-.queue-ad-fallback-gradient {
-  position: absolute;
-  inset: 0;
-  background:
-    radial-gradient(circle at top right, rgba(118, 185, 0, 0.22), transparent 42%),
-    linear-gradient(180deg, rgba(2, 6, 4, 0.2), rgba(2, 6, 4, 0.82));
-}
-
-.queue-ad-fallback-grid {
-  position: absolute;
-  inset: 0;
-  opacity: 0.12;
-  background-image:
-    linear-gradient(rgba(255, 255, 255, 0.18) 1px, transparent 1px),
-    linear-gradient(90deg, rgba(255, 255, 255, 0.18) 1px, transparent 1px);
-  background-size: 28px 28px;
-  mask-image: linear-gradient(180deg, rgba(0, 0, 0, 0.9), transparent 100%);
-}
-
-.queue-ad-fallback-content {
-  position: relative;
-  z-index: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 18px;
-  min-height: 180px;
-  padding: 18px;
-}
-
-.queue-ad-fallback--compact .queue-ad-fallback-content {
-  min-height: 160px;
-  padding: 16px;
-}
-
-.queue-ad-fallback-badge {
-  display: inline-flex;
-  width: fit-content;
-  align-items: center;
-  gap: 8px;
-  padding: 6px 10px;
-  border-radius: 999px;
-  background: rgba(118, 185, 0, 0.14);
-  color: #c9ff6a;
-  font-size: 0.74rem;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-}
-
-.queue-ad-fallback-copy {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  max-width: 420px;
-}
-
-.queue-ad-fallback-title {
-  margin: 0;
-  color: #fff;
-  font-size: 1.05rem;
-  font-weight: 800;
-  line-height: 1.15;
-}
-
-.queue-ad-fallback--compact .queue-ad-fallback-title {
-  font-size: 0.95rem;
-}
-
-.queue-ad-fallback-message {
-  margin: 0;
-  color: rgba(255, 255, 255, 0.82);
-  font-size: 0.82rem;
-  line-height: 1.5;
-}
-
 .queue-ad-preview-frame {
   position: relative;
   overflow: hidden;
