+++ opennow-stable/src/shared/gfn/auth.ts
@@ -1,0 +1,122 @@
+export interface LoginProvider {
+  idpId: string;
+  code: string;
+  displayName: string;
+  streamingServiceUrl: string;
+  priority: number;
+}
+
+export interface AuthTokens {
+  accessToken: string;
+  refreshToken?: string;
+  idToken?: string;
+  expiresAt: number;
+  clientToken?: string;
+  clientTokenExpiresAt?: number;
+  clientTokenLifetimeMs?: number;
+}
+
+export interface AuthUser {
+  userId: string;
+  displayName: string;
+  email?: string;
+  avatarUrl?: string;
+  membershipTier: string;
+}
+
+export interface EntitledResolution {
+  width: number;
+  height: number;
+  fps: number;
+}
+
+export interface StorageAddon {
+  type: "PERMANENT_STORAGE";
+  sizeGb?: number;
+  usedGb?: number;
+  regionName?: string;
+  regionCode?: string;
+}
+
+export interface SubscriptionInfo {
+  membershipTier: string;
+  subscriptionType?: string;
+  subscriptionSubType?: string;
+  allottedHours: number;
+  purchasedHours: number;
+  rolledOverHours: number;
+  usedHours: number;
+  remainingHours: number;
+  totalHours: number;
+  firstEntitlementStartDateTime?: string;
+  serverRegionId?: string;
+  currentSpanStartDateTime?: string;
+  currentSpanEndDateTime?: string;
+  notifyUserWhenTimeRemainingInMinutes?: number;
+  notifyUserOnSessionWhenRemainingTimeInMinutes?: number;
+  state?: string;
+  isGamePlayAllowed?: boolean;
+  isUnlimited: boolean;
+  storageAddon?: StorageAddon;
+  entitledResolutions: EntitledResolution[];
+}
+
+export interface AuthSession {
+  provider: LoginProvider;
+  tokens: AuthTokens;
+  user: AuthUser;
+}
+
+export interface SavedAccount {
+  userId: string;
+  displayName: string;
+  email?: string;
+  avatarUrl?: string;
+  membershipTier: string;
+  providerCode: string;
+}
+
+export interface ThankYouContributor {
+  login: string;
+  avatarUrl: string;
+  profileUrl: string;
+  contributions: number;
+}
+
+export interface ThankYouSupporter {
+  name: string;
+  avatarUrl?: string;
+  profileUrl?: string;
+  isPrivate: boolean;
+}
+
+export interface ThankYouDataResult {
+  contributors: ThankYouContributor[];
+  supporters: ThankYouSupporter[];
+  contributorsError?: string;
+  supportersError?: string;
+}
+
+export interface AuthLoginRequest {
+  providerIdpId?: string;
+}
+
+export interface AuthSessionRequest {
+  forceRefresh?: boolean;
+}
+
+export type AuthRefreshOutcome = "not_attempted" | "refreshed" | "failed" | "missing_refresh_token";
+
+export interface AuthRefreshStatus {
+  attempted: boolean;
+  forced: boolean;
+  outcome: AuthRefreshOutcome;
+  message: string;
+  error?: string;
+}
+
+export interface AuthSessionResult {
+  session: AuthSession | null;
+  refresh: AuthRefreshStatus;
+}
+
